[MIG] pos_ux: Migration to 17.0
X-original-commit: 40058f90ed3c150b92745d4bae1ae8ab833f9fc8
