+++ pos_ux/__manifest__.py
@@ -1,6 +1,6 @@
 {
     'name': 'Point of Sale UX',
-    'version': '16.0.1.0.0',
+    'version': "17.0.1.0.0",
     'category': 'Point of Sale',
     'description': """
 This module extend functionality of point of sale .
@@ -13,7 +13,7 @@ This module extend functionality of point of sale .
         'views/res_config_settings_view.xml',
     ],
     'assets': {
-        'point_of_sale.assets': [
+        'point_of_sale._assets_pos': [
             'pos_ux/static/src/**/*'
         ],
     },
