+++ pos_ux/static/src/js/payment_screen/payment_screen.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<templates id="template" xml:space="preserve">
+    <t t-name="PaymentScreen" t-inherit="point_of_sale.PaymentScreenButtons" t-inherit-mode="extension">
+        <xpath expr="//button[hasclass('js_invoice')]" position="attributes">
+            <attribute name="t-if">pos.config.billing_behavior != 'invoice_required'</attribute>
+        </xpath>
+    </t>
+</templates>
