--- pos_ux/static/src/xml/Chrome.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates id="template" xml:space="preserve">
-
-    <t t-name="Chrome" t-inherit="point_of_sale.Chrome" t-inherit-mode="extension" owl="1">
-        <xpath expr="//HeaderButton" position="before">
-        <div class="header-button invoice-contingency"
-             t-on-click="env.pos.setContingencyMode">
-            <span class="contingency-button ">
-                <i class="fa"
-                   t-att-class="{'fa-unlink': env.pos.pos_session.invoice_contingency, 'fa-link': !env.pos.pos_session.invoice_contingency }"
-                   role="img" aria-label="Contingency mode" title="Contingency mode"></i>
-            </span>
-        </div>
-
-        </xpath>
-    </t>
-
-</templates>
