--- pos_ux/static/src/xml/PaymentScreen.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates id="template" xml:space="preserve">
-    <t t-name="PaymentScreen" t-inherit="point_of_sale.PaymentScreen"  owl="1">
-        <xpath expr="//div[hasclass('js_invoice')]" position="attributes">
-            <attribute name="t-if">env.pos.config.billing_behavior != 'invoice_required'</attribute>
-        </xpath>
-    </t>
-</templates>
